[Refactor] Move storage iterator primitives into StoragePrimitive
