+++ be/src/storage/aggregate_iterator.h
@@ -14,8 +14,7 @@
 
 #pragma once
 
-#include "storage/chunk_iterator.h"
-#include "storage/row_source_mask.h"
+#include "storage/primitive/chunk_iterator.h"
 
 namespace starrocks {
