+++ be/src/storage/chunk_aggregator.h
@@ -20,9 +20,9 @@
 #include "column/chunk.h"
 #include "column/vectorized_fwd.h"
 #include "storage/chunk_helper.h"
-#include "storage/chunk_iterator.h"
 #include "storage/column_aggregate_func.h"
-#include "storage/row_source_mask.h"
+#include "storage/primitive/chunk_iterator.h"
+#include "storage/primitive/row_source_mask.h"
 
 namespace starrocks {
