+++ be/src/storage/empty_iterator.h
@@ -14,11 +14,4 @@
 
 #pragma once
 
-#include "storage/chunk_iterator.h"
-
-namespace starrocks {
-
-ChunkIteratorPtr new_empty_iterator(Schema&& schema, int chunk_size);
-ChunkIteratorPtr new_empty_iterator(const Schema& schema, int chunk_size);
-
-} // namespace starrocks
+#include "storage/primitive/empty_iterator.h"
