+++ be/src/storage/lake/tablet_reader.h
@@ -16,9 +16,9 @@
 
 #include "exec/pipeline/scan/scan_morsel.h"
 #include "runtime/mem_pool.h"
-#include "storage/chunk_iterator.h"
 #include "storage/delete_predicates.h"
 #include "storage/lake/versioned_tablet.h"
+#include "storage/primitive/chunk_iterator.h"
 #include "storage/tablet_reader_params.h"
 #include "types_fwd.h"
