+++ be/src/storage/lake/update_compaction_state.cpp
@@ -20,10 +20,10 @@
 #include "gutil/strings/substitute.h"
 #include "runtime/current_thread.h"
 #include "storage/chunk_helper.h"
-#include "storage/chunk_iterator.h"
 #include "storage/lake/rowset.h"
 #include "storage/lake/update_manager.h"
 #include "storage/primary_key_encoder.h"
+#include "storage/primitive/chunk_iterator.h"
 #include "storage/tablet_manager.h"
 
 namespace starrocks::lake {
