+++ be/src/storage/merge_iterator.h
@@ -16,7 +16,7 @@
 
 #include <vector>
 
-#include "storage/chunk_iterator.h"
+#include "storage/primitive/chunk_iterator.h"
 #include "storage/row_source_mask.h"
 
 namespace starrocks {
