+++ be/src/storage/primary_key_compaction_conflict_resolver.h
@@ -18,9 +18,9 @@
 
 #include "common/status.h"
 #include "fs/fs.h"
-#include "storage/chunk_iterator.h"
 #include "storage/del_vector.h"
 #include "storage/primary_key_encoding_types.h"
+#include "storage/primitive/chunk_iterator.h"
 
 namespace starrocks {
