+++ be/src/storage/primary_key_dump.cpp
@@ -23,10 +23,10 @@
 #include "fs/fs_factory.h"
 #include "fs/fs_util.h"
 #include "storage/chunk_helper.h"
-#include "storage/chunk_iterator.h"
 #include "storage/del_vector.h"
 #include "storage/delta_column_group.h"
 #include "storage/primary_key_encoder.h"
+#include "storage/primitive/chunk_iterator.h"
 #include "storage/rowset/rowset.h"
 #include "storage/rowset/segment_iterator.h"
 #include "storage/rowset/segment_options.h"
