+++ be/src/storage/primary_key_dump.h
@@ -18,8 +18,8 @@
 #include <unordered_map>
 
 #include "common/status.h"
-#include "storage/chunk_iterator.h"
 #include "storage/olap_common.h"
+#include "storage/primitive/chunk_iterator.h"
 #include "storage/tablet_schema.h"
 
 namespace starrocks {
