+++ be/src/storage/primary_key_recover.h
@@ -19,8 +19,8 @@
 
 #include "column/schema.h"
 #include "common/status.h"
-#include "storage/chunk_iterator.h"
 #include "storage/primary_index.h"
+#include "storage/primitive/chunk_iterator.h"
 
 namespace starrocks {
