+++ be/src/storage/primitive/chunk_iterator.cpp
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "storage/chunk_iterator.h"
+#include "storage/primitive/chunk_iterator.h"
 
 namespace starrocks {
