+++ be/src/storage/primitive/chunk_iterator.h
@@ -15,13 +15,15 @@
 #pragma once
 
 #include <memory>
+#include <unordered_set>
+#include <vector>
 
 #include "column/chunk.h"
 #include "column/schema.h"
 #include "common/constexpr.h"
 #include "common/runtime_profile.h"
 #include "runtime/global_dict/types.h"
-#include "storage/row_source_mask.h"
+#include "storage/primitive/row_source_mask.h"
 
 namespace starrocks {
 class Status;
