+++ be/src/storage/primitive/projection_iterator.h
@@ -18,8 +18,12 @@
 
 namespace starrocks {
 
-// A unique iterator will removes all but the *last* record from every
-// consecutive group of records with equivalent keys.
-ChunkIteratorPtr new_unique_iterator(const ChunkIteratorPtr& child);
+// A projection iterator will fetch chunk from its |child| iterator and choosing the columns
+// specified in |schema|.
+// it is undefined if a column specified in |schema| not exists in the |child|'s schema.
+// the relative order of columns in |schema| can be different from that in |child|'s schema, e.g,
+// |child| may return columns in the order `c1`, `c2`, `c3`, the projection iterator may return
+// columns in the order `c3`, `c1`.
+ChunkIteratorPtr new_projection_iterator(const Schema& schema, const ChunkIteratorPtr& child);
 
-} // namespace starrocks
+} //  namespace starrocks
