+++ be/src/storage/rowset/rowset.cpp
@@ -51,12 +51,12 @@
 #include "runtime/runtime_state.h"
 #include "segment_options.h"
 #include "storage/chunk_helper.h"
-#include "storage/chunk_iterator.h"
 #include "storage/delete_predicates.h"
-#include "storage/empty_iterator.h"
 #include "storage/index/index_descriptor.h"
 #include "storage/merge_iterator.h"
-#include "storage/projection_iterator.h"
+#include "storage/primitive/chunk_iterator.h"
+#include "storage/primitive/empty_iterator.h"
+#include "storage/primitive/projection_iterator.h"
 #include "storage/rowset/metadata_cache.h"
 #include "storage/rowset/rowid_range_option.h"
 #include "storage/rowset/short_key_range_option.h"
