+++ be/src/storage/rowset/rowset_writer.cpp
@@ -60,11 +60,11 @@
 #include "serde/column_array_serde.h"
 #include "storage/aggregate_iterator.h"
 #include "storage/chunk_helper.h"
-#include "storage/empty_iterator.h"
 #include "storage/index/index_descriptor.h"
 #include "storage/merge_iterator.h"
 #include "storage/metadata_util.h"
 #include "storage/olap_define.h"
+#include "storage/primitive/empty_iterator.h"
 #include "storage/row_source_mask.h"
 #include "storage/rows_mapper.h"
 #include "storage/rowset/rowset.h"
