+++ be/src/storage/rowset/segment_iterator.cpp
@@ -46,7 +46,6 @@
 #include "runtime/chunk_helper.h"
 #include "segment_options.h"
 #include "storage/chunk_helper.h"
-#include "storage/chunk_iterator.h"
 #include "storage/column_expr_predicate.h"
 #include "storage/column_or_predicate.h"
 #include "storage/column_predicate.h"
@@ -61,9 +60,10 @@
 #include "storage/index/vector/vector_search_option.h"
 #include "storage/lake/filenames.h"
 #include "storage/lake/update_manager.h"
+#include "storage/primitive/chunk_iterator.h"
+#include "storage/primitive/projection_iterator.h"
 #include "storage/primitive/range.h"
 #include "storage/primitive/rowid_types.h"
-#include "storage/projection_iterator.h"
 #include "storage/roaring2range.h"
 #include "storage/rowset/bitmap_index_evaluator.h"
 #include "storage/rowset/bitmap_index_reader.h"
