+++ be/src/storage/rowset_merger.cpp
@@ -27,9 +27,9 @@
 #include "common/config_exec_fwd.h"
 #include "gutil/stl_util.h"
 #include "storage/chunk_helper.h"
-#include "storage/empty_iterator.h"
 #include "storage/merge_iterator.h"
 #include "storage/primary_key_encoder.h"
+#include "storage/primitive/empty_iterator.h"
 #include "storage/rowset/column_reader.h"
 #include "storage/rowset/rowset_options.h"
 #include "storage/rowset/rowset_writer.h"
