+++ be/src/storage/table_reader.h
@@ -19,8 +19,8 @@
 
 #include "common/brpc/internal_service_recoverable_stub.h"
 #include "gen_cpp/Descriptors_types.h"
-#include "storage/chunk_iterator.h"
 #include "storage/column_predicate.h"
+#include "storage/primitive/chunk_iterator.h"
 #include "storage/tablet.h"
 
 namespace starrocks {
