+++ be/src/storage/tablet_reader.cpp
@@ -35,10 +35,10 @@
 #include "storage/column_predicate_rewriter.h"
 #include "storage/conjunctive_predicates.h"
 #include "storage/delete_predicates.h"
-#include "storage/empty_iterator.h"
 #include "storage/merge_iterator.h"
 #include "storage/olap_common.h"
 #include "storage/predicate_parser.h"
+#include "storage/primitive/empty_iterator.h"
 #include "storage/rowset/column_reader.h"
 #include "storage/rowset/rowid_range_option.h"
 #include "storage/seek_range.h"
