+++ be/src/storage/tablet_reader_params.h
@@ -21,9 +21,9 @@
 #include "column/column_access_path.h"
 #include "options.h"
 #include "runtime/global_dict/types.h"
-#include "storage/chunk_iterator.h"
 #include "storage/olap_common.h"
 #include "storage/predicate_tree/predicate_tree.hpp"
+#include "storage/primitive/chunk_iterator.h"
 #include "storage/primitive/olap_tuple.h"
 #include "storage/runtime_filter_predicate.h"
 #include "storage/runtime_range_pruner.h"
