+++ be/test/storage/get_use_pk_index_test.cpp
@@ -29,9 +29,9 @@
 #include "gutil/walltime.h"
 #include "runtime/runtime_state.h"
 #include "storage/chunk_helper.h"
-#include "storage/empty_iterator.h"
 #include "storage/kv_store.h"
 #include "storage/primary_key_encoder.h"
+#include "storage/primitive/empty_iterator.h"
 #include "storage/rowset/rowset_factory.h"
 #include "storage/rowset/rowset_writer.h"
 #include "storage/rowset/rowset_writer_context.h"
