+++ be/test/storage/publish_version_manager_test.cpp
@@ -29,8 +29,8 @@
 #include "gtest/gtest.h"
 #include "storage/chunk_helper.h"
 #include "storage/delta_writer.h"
-#include "storage/empty_iterator.h"
 #include "storage/options.h"
+#include "storage/primitive/empty_iterator.h"
 #include "storage/rowset/rowset_factory.h"
 #include "storage/rowset/rowset_writer.h"
 #include "storage/rowset/rowset_writer_context.h"
