+++ be/test/storage/rowset/flat_json_column_rw_test.cpp
@@ -35,9 +35,9 @@
 #include "gutil/casts.h"
 #include "storage/aggregate_type.h"
 #include "storage/chunk_helper.h"
-#include "storage/chunk_iterator.h"
 #include "storage/flat_json_config.h"
 #include "storage/olap_common.h"
+#include "storage/primitive/chunk_iterator.h"
 #include "storage/rowset/column_iterator.h"
 #include "storage/rowset/column_reader.h"
 #include "storage/rowset/column_writer.h"
