+++ be/test/storage/rowset/rowset_test.cpp
@@ -54,8 +54,8 @@
 #include "runtime/mem_pool.h"
 #include "runtime/mem_tracker.h"
 #include "storage/chunk_helper.h"
-#include "storage/chunk_iterator.h"
-#include "storage/empty_iterator.h"
+#include "storage/primitive/chunk_iterator.h"
+#include "storage/primitive/empty_iterator.h"
 #include "storage/rowset/rowset_factory.h"
 #include "storage/rowset/rowset_options.h"
 #include "storage/rowset/rowset_writer.h"
