+++ be/test/storage/rowset_column_update_state_test.cpp
@@ -27,8 +27,8 @@
 #include "runtime/mem_pool.h"
 #include "runtime/mem_tracker.h"
 #include "storage/chunk_helper.h"
-#include "storage/empty_iterator.h"
 #include "storage/olap_common.h"
+#include "storage/primitive/empty_iterator.h"
 #include "storage/rowset/rowset_factory.h"
 #include "storage/rowset/rowset_options.h"
 #include "storage/storage_engine.h"
