+++ be/test/storage/tablet_updates_test.h
@@ -32,9 +32,9 @@
 #include "gutil/strings/substitute.h"
 #include "gutil/walltime.h"
 #include "storage/chunk_helper.h"
-#include "storage/empty_iterator.h"
 #include "storage/kv_store.h"
 #include "storage/primary_key_encoder.h"
+#include "storage/primitive/empty_iterator.h"
 #include "storage/rowset/rowset_factory.h"
 #include "storage/rowset/rowset_options.h"
 #include "storage/rowset/rowset_writer.h"
