+++ be/test/storage/vector_chunk_iterator.h
@@ -21,7 +21,7 @@
 #include "column/chunk.h"
 #include "column/datum_convert.h"
 #include "storage/chunk_helper.h"
-#include "storage/chunk_iterator.h"
+#include "storage/primitive/chunk_iterator.h"
 #include "storage/types.h"
 #include "types/datum.h"
 #include "types/storage_type_traits.h"
